+++ fontes/completude/pitugues-provedor-completude.ts
@@ -5,18 +5,13 @@ import { obterResultado } from '../analise-codigo/cache-analise';
 import { ParametroDetectado, TipoParametro } from '../interfaces/completude';
 
 import { Classe, FuncaoDeclaracao } from '@designliquido/delegua/declaracoes';
-import primitivasDicionario from '@designliquido/delegua/bibliotecas/dialetos/pitugues/primitivas-dicionario';
-import primitivasNumero from '@designliquido/delegua/bibliotecas/dialetos/pitugues/primitivas-numero';
-import primitivasTexto from '@designliquido/delegua/bibliotecas/dialetos/pitugues/primitivas-texto';
-import primitivasVetor from '@designliquido/delegua/bibliotecas/dialetos/pitugues/primitivas-vetor';
-import {funcoesNativasPitugues} from '../bibliotecas/dialetos/pitugues/funcoes-nativas';
-import { formatarPrimitivas } from '../bibliotecas';
-
-
-const primitivasDicionarioFormatadas = formatarPrimitivas(primitivasDicionario);
-const primitivasNumeroFormatadas = formatarPrimitivas(primitivasNumero);
-const primitivasTextoFormatadas = formatarPrimitivas(primitivasTexto);
-const primitivasVetorFormatadas = formatarPrimitivas(primitivasVetor);
+import {
+    primitivasDicionarioFormatadas,
+    primitivasNumeroFormatadas,
+    primitivasTextoFormatadas,
+    primitivasVetorFormatadas,
+    funcoesNativasPitugues,
+} from '../bibliotecas/dialetos/pitugues';
 
 /**
  * Classe de provedor de completude de Pituguês.
@@ -404,7 +399,7 @@ export class PituguesProvedorCompletude implements vscode.CompletionItemProvider
                             return itemCompletude;
                         });
                     default:
-                        return [];
+                        break;
                 }
             }
         }